Implemented Notify for all prev. suspended orders and now completed.

+++ src/Messages/Responses/OrderResponse.java
@@ -1,13 +1,12 @@
 package Messages.Responses;
 
 import Exchange.Order;
-import Exchange.OrderType;
 import Messages.CrossMessage;
 import Messages.OperationType;
 
 public abstract class OrderResponse extends CrossMessage
 {
-    private final int orderID;
+    private final long orderID;
 
     protected OrderResponse(OperationType type, Order order)
     {
@@ -15,7 +14,7 @@ public abstract class OrderResponse extends CrossMessage
         this.orderID = order.GetOrderID();
     }
 
-    public int GetOrderID()
+    public long GetOrderID()
     {
         return orderID;
     }
